requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.0
backcall==0.2.0
black==21.4b2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
detectron2==0.6
distro==1.8.0
docker-pycreds==0.4.0
exceptiongroup==1.2.0
executing==2.0.0
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.17.2
hydra-core==1.3.2
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
info-nce-pytorch==0.1.4
iopath==0.1.9
ipykernel==6.25.2
ipython==8.12.3
jedi==0.19.1
jupyter-client==8.3.1
jupyter-core==5.3.2
kiwisolver==1.4.5
lazy-loader==0.3
loralib==0.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.22.4
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.16.0
openai==1.3.4
opencv-python==4.8.0.76
optim==0.1.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pbd==0.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.5.1
pydantic-core==2.14.3
pydot==1.4.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
spark-parser==1.8.9
stack-data==0.6.3
submitit==1.4.6
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tifffile==2023.7.10
timm==0.9.7
toml==0.10.2
torch==1.12.1+cu113
torch-summary==1.4.5
torchaudio==0.12.1+cu113
torchinfo==1.8.0
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing-extensions==4.8.0
uncompyle6==3.9.0
uritemplate==4.1.1
urllib3==1.26.16
wandb==0.15.11
wcwidth==0.2.6
werkzeug==2.3.7
xdis==6.0.5
yacs==0.1.8
zipp==3.17.0
